requirements.txt
alembic==1.11.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
filelock==3.12.2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
platformdirs==3.9.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
virtualenv==20.24.0
wcwidth==0.2.6
Werkzeug==2.3.6
